#528 supervisor configurator cleanup
