+++ supervisor/startup.lua
@@ -22,7 +22,7 @@ local supervisor = require("supervisor.supervisor")
 
 local svsessions = require("supervisor.session.svsessions")
 
-local SUPERVISOR_VERSION = "v1.5.6"
+local SUPERVISOR_VERSION = "v1.5.7"
 
 local println = util.println
 local println_ts = util.println_ts
